Removed un-needed components. Moved to using more TS. Updated style throughout.
