+++ App.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+import NaviagtionTabs from "./app/screens/NavigationTabs";
+
+function App() {
+    return <NaviagtionTabs />;
+}
+
+export default App;
