--- app/components/Chart.js
@@ -1,5 +1,0 @@
-function Chart({ type, obj }) {
-    return type(obj);
-}
-
-export default Chart;
