--- app/components/forms/AppForm.js
@@ -1,16 +1,0 @@
-import React from "react";
-import { Formik } from "formik";
-
-function AppForm({ initialValues, onSubmit, validationSchema, children }) {
-    return (
-        <Formik
-            initialValues={initialValues}
-            onSubmit={onSubmit}
-            validationSchema={validationSchema}
-        >
-            {() => <>{children}</>}
-        </Formik>
-    );
-}
-
-export default AppForm;
