--- app/components/forms/AppFormField.js
@@ -1,24 +1,0 @@
-import React from "react";
-import { useFormikContext } from "formik";
-
-import AppTextInput from "../AppTextInput";
-import ErrorMessage from "./ErrorMessage";
-
-function AppFormField({ name, width, ...otherProps }) {
-    const { setFieldTouched, handleChange, errors, touched } =
-        useFormikContext();
-
-    return (
-        <>
-            <AppTextInput
-                onBlur={() => setFieldTouched(name)}
-                onChangeText={handleChange(name)}
-                width={width}
-                {...otherProps}
-            />
-            <ErrorMessage error={errors[name]} visible={touched[name]} />
-        </>
-    );
-}
-
-export default AppFormField;
