--- app/components/forms/AppFormPicker.js
@@ -1,23 +1,0 @@
-import React from "react";
-import { useFormikContext } from "formik";
-
-import AppPicker from "../AppPicker";
-import ErrorMessage from "./ErrorMessage";
-
-function AppFormPicker({ items, name, placeholder, width }) {
-    const { setFieldValue, values } = useFormikContext();
-
-    return (
-        <>
-            <AppPicker
-                items={items}
-                onSelectItem={(item) => setFieldValue(name, item)}
-                placeholder={placeholder}
-                selectedItem={values[name]}
-                width={width}
-            />
-        </>
-    );
-}
-
-export default AppFormPicker;
