--- app/components/forms/index.js
@@ -1,5 +1,0 @@
-export { default as AppForm } from "./AppForm";
-export { default as AppFormField } from "./AppFormField";
-export { default as AppFormPicker } from "./AppFormPicker";
-export { default as ErrorMessage } from "./ErrorMessage";
-export { default as SubmitButton } from "./SubmitButton";
