+++ app/config/colors.js
@@ -1,11 +1,12 @@
 export default {
+    black: "#000000",
+    white: "#ffffff",
+    lightGray: "#EFEFEF",
     primary: "#002657",
     secondary: "#19212C",
-    black: "#000",
-    white: "#fff",
     medium: "#6e6969",
     light: "#f8f4f4",
-    dark: "#0c0c0c",
+    dark: "#1c1c1c",
     danger: "#ff5252",
     dark_blue: "#19212C",
 };
