--- app/config/counties.js
@@ -1,13 +1,0 @@
-export default counties = [
-    { label: "Sedgwick", value: 1 },
-    { label: "Sedgwick", value: 2 },
-    { label: "Sedgwick", value: 3 },
-    { label: "Sedgwick", value: 4 },
-    { label: "Sedgwick", value: 5 },
-    { label: "Sedgwick", value: 5 },
-    { label: "Sedgwick", value: 6 },
-    { label: "Sedgwick", value: 7 },
-    { label: "Sedgwick", value: 8 },
-    { label: "Sedgwick", value: 9 },
-    { label: "Sedgwick", value: 10 },
-];
