--- app/config/graph.js
@@ -1,22 +1,0 @@
-export default graphs = [
-    {
-        id: 1,
-        title: "Salary",
-        details: "Total Annual Salary for each United States county.",
-        image: require("../assets/2020_salary.png"),
-    },
-    {
-        id: 2,
-        title: "Temperature Average",
-        details:
-            "Avearage Temperagre across the United States per County for 2020.",
-        image: require("../assets/2020_tavg.png"),
-    },
-    {
-        id: 3,
-        title: "T Avg vs Salary Correlation",
-        details:
-            "Avearage Temperagre across the United States per County for 2020.",
-        image: require("../assets/tavg_pvalue.png"),
-    },
-];
