--- app/config/listing.js
@@ -1,31 +1,0 @@
-var pcp_pic = require("../assets/2020_pcp.png");
-var sal_pic = require("../assets/2020_salary.png");
-
-export default listings = [
-    {
-        id: 1,
-        title: "Precipitation",
-        details: "This is precipation data pulled from ",
-    },
-    {
-        id: 2,
-        title: "Salary",
-        details: "Total Annual Salary for each United States county.",
-    },
-    {
-        id: 3,
-        title: "Temperature Average",
-        details:
-            "Avearage Temperagre across the United States per County for 2020.",
-    },
-    {
-        id: 4,
-        title: "Severe Weather",
-        details: "This is stuff about stuff",
-    },
-    {
-        id: 5,
-        title: "Wild Fire",
-        details: "This is data that was pulled form ",
-    },
-];
