+++ app/screens/ChartScreen.tsx
@@ -7,10 +7,9 @@ import {
 } from "react-native";
 
 import Card from "../components/Card";
-import Chart from "../components/Chart";
+import { Chart } from "../components/Chart";
 import { ChartApi } from "../apis/ChartApi";
-import ChartModel from "../models/ChartModel";
-import ChoroplethMapModel from "../models/ChoroplethMapModel";
+import * as Model from "../models/Chart";
 import colors from "../config/colors";
 import Screen from "../components/Screen";
 
@@ -18,15 +17,15 @@ interface ChartScreenProps {
     chartIds: Array<number>;
 }
 
-function ChartScreen({ chartIds }: ChartScreenProps) {
-    const [charts, setCharts] = useState<Array<ChartModel>>([]);
+export function ChartScreen({ chartIds }: ChartScreenProps) {
+    const [charts, setCharts] = useState<Array<Model.Chart>>([]);
 
     function mapChartResponseToModel(
         chartResponse: any
-    ): ChartModel | undefined {
+    ): Model.Chart | undefined {
         switch (chartResponse.type) {
             case "choropleth_map":
-                return ChoroplethMapModel.mapResponse(chartResponse);
+                return Model.ChoroplethMap.mapResponse(chartResponse);
             default:
                 console.error(
                     "Could not find valid type to map response type to."
@@ -37,9 +36,9 @@ function ChartScreen({ chartIds }: ChartScreenProps) {
 
     useEffect(() => {
         async function getCharts(): Promise<void> {
-            let chartsFromApi: Array<ChartModel> = [];
+            let chartsFromApi: Array<Model.Chart> = [];
             for await (const chartId of chartIds) {
-                let response: ChartModel = await ChartApi.getChart(chartId);
+                let response: Model.Chart = await ChartApi.getChart(chartId);
 
                 let chart = mapChartResponseToModel(response!);
                 if (chart && !chartsFromApi.includes(chart)) {
@@ -54,49 +53,33 @@ function ChartScreen({ chartIds }: ChartScreenProps) {
 
     return (
         <Screen style={styles.screen}>
-            <ImageBackground
-                source={require("../assets/background.png")}
-                style={styles.image}
-            >
-                <TouchableWithoutFeedback>
-                    <FlatList
-                        data={charts}
-                        keyExtractor={(item) => item.id.toString()}
-                        renderItem={({ item }) => {
-                            let chart = <Chart type={item.type} obj={item} />;
-                            return (
-                                <Card
-                                    title={item.title}
-                                    subtitle={item.subtitle}
-                                    inner={chart}
-                                />
-                            );
-                        }} />
-                </TouchableWithoutFeedback>
-            </ImageBackground>
+            <FlatList
+                data={charts}
+                keyExtractor={(item) => item.id.toString()}
+                renderItem={({ item }) => {
+                    let chart = <Chart type={item.type} obj={item} />;
+                    return (
+                        <Card
+                            title={item.title}
+                            subtitle={item.subtitle}
+                            inner={chart}
+                        />
+                    );
+                }}
+            />
         </Screen>
     );
 }
 
 const styles = StyleSheet.create({
     screen: {
-        backgroundColor: colors.primary,
-    },
-    image: {
-        flex: 1,
-        justifyContent: "flex-end",
+        backgroundColor: colors.white,
+        borderTopRightRadius: 15,
+        borderTopLeftRadius: 15,
     },
     chartRow: {
         flex: 1,
         width: "100%",
-    },
-    container: {
-        paddingTop: 30,
-        width: "100%",
-        height: "100%",
-        backgroundColor: "#fff",
-        alignItems: "center",
-        justifyContent: "center",
     },
 });
