+++ package.json
@@ -27,8 +27,10 @@
         "web": "expo start --web"
     },
     "dependencies": {
+        "@react-native-community/checkbox": "^0.5.14",
         "@react-navigation/bottom-tabs": "^6.4.0",
         "@react-navigation/native": "^6.0.13",
+        "@types/d3": "^7.4.0",
         "axios": "^1.1.3",
         "d3": "^7.6.1",
         "expo": "~46.0.13",
@@ -38,7 +40,8 @@
         "react-native-inset-shadow": "^1.0.3",
         "react-native-plotly": "^6.0.0",
         "react-native-reanimated": "~2.9.1",
-        "reanimated-collapsible-helpers": "^1.0.0"
+        "reanimated-collapsible-helpers": "^1.0.0",
+        "expo-checkbox": "~2.2.0"
     },
     "devDependencies": {
         "@babel/preset-typescript": "^7.18.6",
